+++ golang/cmd/mqtt-to-postgresql/dataprocessing.go
@@ -10,235 +10,6 @@ import (
 	"go.uber.org/zap"
 )
 
-type stateQueue struct {
-	DBAssetID   uint32
-	State       uint32
-	TimestampMs uint64
-}
-type state struct {
-	State       uint32 `json:"state"`
-	TimestampMs uint64 `json:"timestamp_ms"`
-}
-
-// ProcessStateData processes an incoming state message
-func ProcessStateData(customerID string, location string, assetID string, payloadType string, payload []byte, pg *goque.PriorityQueue) error {
-
-	var parsedPayload state
-
-	err := json.Unmarshal(payload, &parsedPayload)
-	if err != nil {
-		zap.S().Errorf("json.Unmarshal failed", err, payload)
-		return nil
-	}
-
-	DBassetID := GetAssetID(customerID, location, assetID)
-
-	newObject := stateQueue{
-		TimestampMs: parsedPayload.TimestampMs,
-		State:       parsedPayload.State,
-		DBAssetID:   DBassetID,
-	}
-
-	marshal, err := json.Marshal(newObject)
-	if err != nil {
-		return nil
-	}
-
-	err = addNewItemToQueue(pg, Prefix.State, marshal)
-	if err != nil {
-		zap.S().Errorf("Error enqueuing", err)
-		return nil
-	}
-
-	return nil
-}
-
-type countQueue struct {
-	DBAssetID   uint32
-	Count       uint32
-	Scrap       uint32
-	TimestampMs uint64
-}
-type count struct {
-	Count       uint32 `json:"count"`
-	Scrap       uint32 `json:"scrap"`
-	TimestampMs uint64 `json:"timestamp_ms"`
-}
-
-// ProcessCountData processes an incoming count message
-func ProcessCountData(customerID string, location string, assetID string, payloadType string, payload []byte, pg *goque.PriorityQueue) error {
-
-	var parsedPayload count
-
-	err := json.Unmarshal(payload, &parsedPayload)
-	if err != nil {
-		zap.S().Errorf("json.Unmarshal failed", err, payload)
-		return nil
-	}
-
-	// this should not happen. Throw a warning message and ignore (= do not try to store in database)
-	if parsedPayload.Count <= 0 {
-		zap.S().Warnf("count <= 0", customerID, location, assetID, payload, parsedPayload)
-		return nil
-	}
-
-	DBassetID := GetAssetID(customerID, location, assetID)
-
-	newObject := countQueue{
-		TimestampMs: parsedPayload.TimestampMs,
-		Count:       parsedPayload.Count,
-		Scrap:       parsedPayload.Scrap,
-		DBAssetID:   DBassetID,
-	}
-
-	marshal, err := json.Marshal(newObject)
-	if err != nil {
-		return nil
-	}
-
-	err = addNewItemToQueue(pg, Prefix.Count, marshal)
-	if err != nil {
-		zap.S().Errorf("Error enqueuing", err)
-		return nil
-	}
-	return nil
-}
-
-type scrapCountQueue struct {
-	DBAssetID   uint32
-	Scrap       uint32
-	TimestampMs uint64
-}
-type scrapCount struct {
-	Scrap       uint32 `json:"scrap"`
-	TimestampMs uint64 `json:"timestamp_ms"`
-}
-
-// ProcessScrapCountData processes an incoming scrapCount message
-func ProcessScrapCountData(customerID string, location string, assetID string, payloadType string, payload []byte, pg *goque.PriorityQueue) error {
-
-	var parsedPayload scrapCount
-
-	err := json.Unmarshal(payload, &parsedPayload)
-	if err != nil {
-		zap.S().Errorf("json.Unmarshal failed", err, payload)
-		return nil
-	}
-
-	DBassetID := GetAssetID(customerID, location, assetID)
-
-	newObject := scrapCountQueue{
-		TimestampMs: parsedPayload.TimestampMs,
-		Scrap:       parsedPayload.Scrap,
-		DBAssetID:   DBassetID,
-	}
-
-	marshal, err := json.Marshal(newObject)
-	if err != nil {
-		return nil
-	}
-
-	err = addNewItemToQueue(pg, Prefix.ScrapCount, marshal)
-	if err != nil {
-		zap.S().Errorf("Error enqueuing", err)
-		return nil
-	}
-	return nil
-}
-
-type addShiftQueue struct {
-	DBAssetID      uint32
-	TimestampMs    uint64
-	TimestampMsEnd uint64
-}
-type addShift struct {
-	TimestampMs    uint64 `json:"timestamp_ms"`
-	TimestampMsEnd uint64 `json:"timestamp_ms_end"`
-}
-
-// ProcessAddShift adds a new shift to the database
-func ProcessAddShift(customerID string, location string, assetID string, payloadType string, payload []byte, pg *goque.PriorityQueue) error {
-
-	var parsedPayload addShift
-
-	err := json.Unmarshal(payload, &parsedPayload)
-	if err != nil {
-		zap.S().Errorf("json.Unmarshal failed", err, payload)
-		return nil
-	}
-
-	DBassetID := GetAssetID(customerID, location, assetID)
-	newObject := addShiftQueue{
-		TimestampMs:    parsedPayload.TimestampMs,
-		TimestampMsEnd: parsedPayload.TimestampMsEnd,
-		DBAssetID:      DBassetID,
-	}
-
-	marshal, err := json.Marshal(newObject)
-	if err != nil {
-		return nil
-	}
-
-	err = addNewItemToQueue(pg, Prefix.AddShift, marshal)
-	if err != nil {
-		zap.S().Errorf("Error enqueuing", err)
-		return nil
-	}
-	return nil
-}
-
-type addMaintenanceActivityQueue struct {
-	DBAssetID     uint32
-	TimestampMs   uint64
-	ComponentName string
-	Activity      int32
-	ComponentID   int32
-}
-type addMaintenanceActivity struct {
-	TimestampMs   uint64 `json:"timestamp_ms"`
-	ComponentName string `json:"component"`
-	Activity      int32  `json:"activity"`
-}
-
-// ProcessAddMaintenanceActivity adds a new maintenance activity to the database
-func ProcessAddMaintenanceActivity(customerID string, location string, assetID string, payloadType string, payload []byte, pg *goque.PriorityQueue) error {
-
-	var parsedPayload addMaintenanceActivity
-
-	err := json.Unmarshal(payload, &parsedPayload)
-	if err != nil {
-		zap.S().Errorf("json.Unmarshal failed", err, payload)
-		return nil
-	}
-
-	DBassetID := GetAssetID(customerID, location, assetID)
-
-	componentID := GetComponentID(DBassetID, parsedPayload.ComponentName)
-	if componentID == 0 {
-		zap.S().Errorf("GetComponentID failed")
-		return nil
-	}
-
-	newObject := addMaintenanceActivityQueue{
-		DBAssetID:     DBassetID,
-		TimestampMs:   parsedPayload.TimestampMs,
-		ComponentName: parsedPayload.ComponentName,
-		ComponentID:   componentID,
-		Activity:      parsedPayload.Activity,
-	}
-	marshal, err := json.Marshal(newObject)
-	if err != nil {
-		return nil
-	}
-
-	err = addNewItemToQueue(pg, Prefix.AddMaintenanceActivity, marshal)
-	if err != nil {
-		zap.S().Errorf("Error enqueuing", err)
-		return nil
-	}
-	return nil
-}
-
 type uniqueProductQueue struct {
 	DBAssetID                  uint32
 	BeginTimestampMs           uint64 `json:"begin_timestamp_ms"`
