+++ golang/cmd/mqtt-to-postgresql/processCountData.go
@@ -1,0 +1,93 @@
+package main
+
+import (
+	"encoding/json"
+	"github.com/beeker1121/goque"
+	"go.uber.org/zap"
+)
+
+type countQueue struct {
+	DBAssetID   uint32
+	Count       uint32
+	Scrap       uint32
+	TimestampMs uint64
+}
+type count struct {
+	Count       uint32 `json:"count"`
+	Scrap       uint32 `json:"scrap"`
+	TimestampMs uint64 `json:"timestamp_ms"`
+}
+
+type CountHandler struct {
+	pg       *goque.PriorityQueue
+	shutdown bool
+}
+
+func (r CountHandler) Setup() (err error) {
+	const queuePathDB = "/data/state/count"
+	r.pg, err = SetupQueue(queuePathDB)
+	if err != nil {
+		zap.S().Errorf("Error setting up remote queue (%s)", queuePathDB, err)
+		return
+	}
+	defer CloseQueue(r.pg)
+	return
+}
+
+func (r CountHandler) process() {
+	for !r.shutdown {
+		//TODO
+	}
+}
+
+func (r CountHandler) enqueue(bytes []byte, priority uint8) {
+	_, err := r.pg.Enqueue(priority, bytes)
+	if err != nil {
+		zap.S().Warnf("Failed to enqueue item", bytes)
+		return
+	}
+}
+
+func (r CountHandler) Shutdown() (err error) {
+	r.shutdown = true
+	err = CloseQueue(r.pg)
+	return
+}
+
+func (r CountHandler) EnqueueMQTT(customerID string, location string, assetID string, payload []byte) {
+
+	var parsedPayload count
+
+	err := json.Unmarshal(payload, &parsedPayload)
+	if err != nil {
+		zap.S().Errorf("json.Unmarshal failed", err, payload)
+		return
+	}
+
+	// this should not happen. Throw a warning message and ignore (= do not try to store in database)
+	if parsedPayload.Count <= 0 {
+		zap.S().Warnf("count <= 0", customerID, location, assetID, payload, parsedPayload)
+		return
+	}
+
+	DBassetID := GetAssetID(customerID, location, assetID)
+
+	newObject := countQueue{
+		TimestampMs: parsedPayload.TimestampMs,
+		Count:       parsedPayload.Count,
+		Scrap:       parsedPayload.Scrap,
+		DBAssetID:   DBassetID,
+	}
+
+	marshal, err := json.Marshal(newObject)
+	if err != nil {
+		return
+	}
+
+	r.enqueue(marshal, 0)
+	if err != nil {
+		zap.S().Errorf("Error enqueuing", err)
+		return
+	}
+	return
+}
