+++ golang/cmd/mqtt-to-postgresql/processScrapCount.go
@@ -1,0 +1,80 @@
+package main
+
+import (
+	"encoding/json"
+	"github.com/beeker1121/goque"
+	"go.uber.org/zap"
+)
+
+type scrapCountQueue struct {
+	DBAssetID   uint32
+	Scrap       uint32
+	TimestampMs uint64
+}
+type scrapCount struct {
+	Scrap       uint32 `json:"scrap"`
+	TimestampMs uint64 `json:"timestamp_ms"`
+}
+
+type ScrapCountHandler struct {
+	pg       *goque.PriorityQueue
+	shutdown bool
+}
+
+func (r ScrapCountHandler) Setup() (err error) {
+	const queuePathDB = "/data/state/ScrapCount"
+	r.pg, err = SetupQueue(queuePathDB)
+	if err != nil {
+		zap.S().Errorf("Error setting up remote queue (%s)", queuePathDB, err)
+		return
+	}
+	defer CloseQueue(r.pg)
+	return
+}
+
+func (r ScrapCountHandler) process() {
+	for !r.shutdown {
+		//TODO
+	}
+}
+
+func (r ScrapCountHandler) enqueue(bytes []byte, priority uint8) {
+	_, err := r.pg.Enqueue(priority, bytes)
+	if err != nil {
+		zap.S().Warnf("Failed to enqueue item", bytes)
+		return
+	}
+}
+
+func (r ScrapCountHandler) Shutdown() (err error) {
+	r.shutdown = true
+	err = CloseQueue(r.pg)
+	return
+}
+
+func (r ScrapCountHandler) EnqueueMQTT(customerID string, location string, assetID string, payload []byte) {
+
+	var parsedPayload scrapCount
+
+	err := json.Unmarshal(payload, &parsedPayload)
+	if err != nil {
+		zap.S().Errorf("json.Unmarshal failed", err, payload)
+		return
+	}
+
+	DBassetID := GetAssetID(customerID, location, assetID)
+
+	newObject := scrapCountQueue{
+		TimestampMs: parsedPayload.TimestampMs,
+		Scrap:       parsedPayload.Scrap,
+		DBAssetID:   DBassetID,
+	}
+
+	marshal, err := json.Marshal(newObject)
+	if err != nil {
+		return
+	}
+
+	r.enqueue(marshal, 0)
+	return
+}
